requirements.txt
# Ask The Storytell AI - Backend Requirements
# FastAPI + Gemini + Stability AI + ElevenLabs

# Backend Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiohttp==3.9.1

# LLM & Embeddings
google-generativeai==0.3.2
openai==1.54.0
sentence-transformers==2.7.0
transformers==4.44.2
huggingface-hub==0.41.1
tokenizers==0.22.0
openai-whisper==20231117

# PDF Processing
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
pydantic==2.5.3

# Optional: Development
# pytest==7.4.3
# black==23.12.1
